Conexion entre gestión de cursos y asignaturas

+++ micro_academico/src/main/java/com/example/micro_academico/gestionCursos/service/CursoService.java
@@ -1,14 +1,15 @@
 package com.example.micro_academico.gestionCursos.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import com.example.micro_academico.gestionCursos.model.entities.Curso;
 import com.example.micro_academico.gestionCursos.repository.CursoRepository;
 import com.example.micro_academico.gestionCursos.model.request.ActualizarCursoRequest;
 import com.example.micro_academico.gestionCursos.model.request.RegistrarCursoRequest;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class CursoService {
@@ -20,8 +21,9 @@ public class CursoService {
         return cursoRepository.findAll();
     }
 
-    public Optional<Curso> obtenerCursoPorId(Long id) {
-        return cursoRepository.findById(id);
+    public Curso obtenerCursoPorId(Long id) {
+        return cursoRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Curso no encontrado"));
     }
 
     public Curso registrarCurso(RegistrarCursoRequest request) {
@@ -35,16 +37,19 @@ public class CursoService {
     }
 
     public Curso actualizarCurso(Long id, ActualizarCursoRequest request) {
-        return cursoRepository.findById(id).map(curso -> {
-            curso.setLetraCurso(request.getLetraCurso());
-            curso.setAnnoAcademico(request.getAnnoAcademico());
-            curso.setJornada(request.getJornada());
-            curso.setCuposMaximos(request.getCuposMaximos());
-            return cursoRepository.save(curso);
-        }).orElseThrow(() -> new RuntimeException("Curso no encontrado con el ID: " + id));
+        Curso curso = obtenerCursoPorId(id); 
+        curso.setLetraCurso(request.getLetraCurso());
+        curso.setAnnoAcademico(request.getAnnoAcademico());
+        curso.setJornada(request.getJornada());
+        curso.setCuposMaximos(request.getCuposMaximos());
+        
+        return cursoRepository.save(curso);
     }
 
     public void eliminarCurso(Long id) {
+        if (!cursoRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Curso no encontrado");
+        }
         cursoRepository.deleteById(id);
     }
 }
